+++ src/main/java/com/deathmotion/marlowcrystal/loader/LoaderAccess.java
@@ -5,6 +5,7 @@ import com.deathmotion.marlowcrystal.loader.fabric.FabricLoaderAccess;
 //?} else {
 /*import com.deathmotion.marlowcrystal.loader.neoforge.NeoForgeLoaderAccess;
 *///?}
+import com.deathmotion.marlowcrystal.update.UpdateResult;
 
 import java.nio.file.Path;
 import java.util.Optional;
@@ -20,4 +21,6 @@ public interface LoaderAccess {
     Path configDirectory();
 
     Optional<String> minecraftVersion();
+
+    void showUpdate(UpdateResult result);
 }